backend/app/utils/image_processing.py
import cv2
import numpy as np
from PIL import Image
import io

def preprocess_image(image_bytes: bytes) -> np.ndarray:
    """
    Preprocess image for CNN model.
    """
    # Convert bytes to PIL Image
    image = Image.open(io.BytesIO(image_bytes))

    # Convert to RGB if necessary
    if image.mode != 'RGB':
        image = image.convert('RGB')

    # Resize to model input size (assuming 224x224)
    image = image.resize((224, 224))

    # Convert to numpy array
    img_array = np.array(image)

    # Normalize
    img_array = img_array / 255.0

    # Add batch dimension
    img_array = np.expand_dims(img_array, axis=0)

    return img_array

def detect_food(image_array: np.ndarray) -> str:
    """
    Placeholder for food detection using CNN model.
    """
    # Load model (placeholder)
    # model = load_model('models/food_cnn_model.h5')
    # prediction = model.predict(image_array)
    # food_class = np.argmax(prediction)

    # For now, return a dummy food
    return "apple"  # Placeholder