backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
scikit-learn==1.3.2
tensorflow==2.15.0
pandas==2.1.4
numpy==1.26.2
opencv-python==4.8.1.78
pillow==10.1.0
matplotlib==3.8.2
plotly==5.17.0
joblib==1.3.2
python-multipart==0.0.6
httpx==0.25.2